Move resample_photoelectron_charge to PhotoelectronSource
