+++ sstcam_simulation/event/acquisition.py
@@ -34,27 +34,6 @@ class EventAcquisition:
         if self.trigger is None:
             self.trigger = NNSuperpixelAboveThreshold(camera)
 
-    def resample_photoelectron_charge(self, pe: Photoelectrons) -> Photoelectrons:
-        """
-        Resample the charges of the photoelectrons from the spectrum defined in
-        the Camera
-
-        Parameters
-        ----------
-        pe : Photoelectrons
-
-        Returns
-        -------
-        Photoelectrons
-        """
-        rng = np.random.default_rng(seed=self.seed)
-        spectrum = self.camera.photoelectron_spectrum
-        n_photoelectrons = pe.pixel.size
-        charge = rng.choice(spectrum.x, size=n_photoelectrons, p=spectrum.pdf)
-        return Photoelectrons(
-            pixel=pe.pixel, time=pe.time, charge=charge, metadata=pe.metadata
-        )
-
     def get_continuous_readout(self, photoelectrons):
         """
         Obtain the sudo-continuous readout from the camera for the given
